refactor: change locations and add packages
