+++ src/main/java/pl/wrapper/parking/facade/dto/stats/parking/ParkingStatsResponse.java
@@ -1,8 +1,8 @@
-package pl.wrapper.parking.facade.dto.stats;
+package pl.wrapper.parking.facade.dto.stats.parking;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import pl.wrapper.parking.facade.dto.stats.basis.ParkingInfo;
-import pl.wrapper.parking.facade.dto.stats.basis.ParkingStats;
+import pl.wrapper.parking.facade.dto.stats.parking.basis.ParkingInfo;
+import pl.wrapper.parking.facade.dto.stats.parking.basis.ParkingStats;
 
 public record ParkingStatsResponse(
         @Schema(implementation = ParkingInfo.class) ParkingInfo parkingInfo,
