+++ src/main/java/pl/wrapper/parking/facade/dto/stats/parking/basis/ParkingStats.java
@@ -1,4 +1,4 @@
-package pl.wrapper.parking.facade.dto.stats.basis;
+package pl.wrapper.parking.facade.dto.stats.parking.basis;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
