+++ src/main/java/pl/wrapper/parking/facade/dto/stats/parking/daily/DailyParkingStatsResponse.java
@@ -1,9 +1,9 @@
-package pl.wrapper.parking.facade.dto.stats.daily;
+package pl.wrapper.parking.facade.dto.stats.parking.daily;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalTime;
-import pl.wrapper.parking.facade.dto.stats.basis.ParkingInfo;
-import pl.wrapper.parking.facade.dto.stats.basis.ParkingStats;
+import pl.wrapper.parking.facade.dto.stats.parking.basis.ParkingInfo;
+import pl.wrapper.parking.facade.dto.stats.parking.basis.ParkingStats;
 
 public record DailyParkingStatsResponse(
         @Schema(implementation = ParkingInfo.class) ParkingInfo parkingInfo,
