+++ src/main/java/pl/wrapper/parking/facade/dto/stats/parking/weekly/CollectiveWeeklyParkingStats.java
@@ -1,12 +1,12 @@
-package pl.wrapper.parking.facade.dto.stats.weekly;
+package pl.wrapper.parking.facade.dto.stats.parking.weekly;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
 import java.util.Map;
 import org.springframework.http.MediaType;
-import pl.wrapper.parking.facade.dto.stats.basis.ParkingInfo;
-import pl.wrapper.parking.facade.dto.stats.basis.ParkingStats;
+import pl.wrapper.parking.facade.dto.stats.parking.basis.ParkingInfo;
+import pl.wrapper.parking.facade.dto.stats.parking.basis.ParkingStats;
 
 public record CollectiveWeeklyParkingStats(
         @Schema(implementation = ParkingInfo.class) ParkingInfo parkingInfo,
